second commit
The initial build out is still largely incomplete, but the state of the
project at this point seems stable enough to be able to move forward
in incremental commits (although admittedly I probably should have done
that from the start).

Note that this repository will likely be destroyed when it is stable
enough to be merged into the PCI Wake server repository.