implement EntityList component
the EntityList component is a generic component to allow for the
configuration of an array of PersistentEntity objects, as is required
for the transient failure strategy's notification strategy.

This introduces a new convention for handling the data binding between
a directive/component and any other directive/component that must be
injected into it using an ng-content directive. Since this seems to be a
successful proof of concept, the rest of the project will be refactored
to use this convention, which will greatly minimize code duplication and
unnecessary couplings.